+++ backend/fit-journey-api/src/main/java/hu/hm/fitjourneyapi/repository/fitness/CardioSetRepository.java
@@ -3,5 +3,5 @@ package hu.hm.fitjourneyapi.repository.fitness;
 import hu.hm.fitjourneyapi.model.fitness.CardioSet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CardioSetRepository extends JpaRepository<CardioSet, Integer> {
+public interface CardioSetRepository extends JpaRepository<CardioSet, Long> {
 }
