+++ backend/fit-journey-api/src/main/java/hu/hm/fitjourneyapi/repository/fitness/SetRepository.java
@@ -3,6 +3,6 @@ package hu.hm.fitjourneyapi.repository.fitness;
 import hu.hm.fitjourneyapi.model.fitness.Set;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SetRepository extends JpaRepository<Set, Integer> {
-    Set getSetById(int id);
+public interface SetRepository extends JpaRepository<Set, Long> {
+    Set getSetById(long id);
 }
